pass the token into the view (#6)
punting on allowing the token to be changed while the app is running for now. basically i want this model to be different than the library model, and it is not yet clear to me at this time in my functional programming career how to see that happen.

so the first time the app is run, it will ask for information to generate the token. it's stored to UserDefaults (and therefore not secure). but by having this mess all in the application, the library will not have to know how the token is generated nor stored. one step at a time.